requirements.txt
selenium>=4.16.0
keybert>=0.7.0
python-dotenv>=1.0.0
textblob>=0.17.1
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain-community>=0.0.16
huggingface-hub>=0.20.3
llama-cpp-python>=0.2.56
transformers>=4.37.0
torch>=2.2.0 